+++ test/unit/infrastructure/random.cpp
@@ -32,16 +32,17 @@ TTS_CASE("Check realistic integral distribution")
   for(auto [ b, v ]: samples)
     input.push_back(v);
 
-  std::vector<int> output(input.size());
+  std::vector<float> output(input.size());
   std::adjacent_difference(input.begin(),
                            input.end(),
                            output.begin(),
-                           [](auto a, auto b) { return std::abs(a) - std::abs(b); });
+                           [](auto a, auto b)
+                           { return static_cast<float>(std::abs(a) - std::abs(b)); });
 
   float average_variation_per_bucket = 0;
   for(std::size_t i = 1; i < output.size() - 1; ++i)
     average_variation_per_bucket += output[ i ];
-  average_variation_per_bucket /= nb;
+  average_variation_per_bucket /= static_cast<float>(nb);
 
   TTS_LESS(average_variation_per_bucket, 0.5f);
 };
@@ -68,16 +69,17 @@ TTS_CASE_TPL("Check realistic real distribution", float, double)
   for(auto [ b, v ]: samples)
     input.push_back(v);
 
-  std::vector<int> output(input.size());
+  std::vector<float> output(input.size());
   std::adjacent_difference(input.begin(),
                            input.end(),
                            output.begin(),
-                           [](auto a, auto b) { return std::abs(a) - std::abs(b); });
+                           [](auto a, auto b)
+                           { return static_cast<float>(std::abs(a) - std::abs(b)); });
 
   float average_variation_per_bucket = 0;
   for(std::size_t i = 1; i < output.size() - 1; ++i)
     average_variation_per_bucket += output[ i ];
-  average_variation_per_bucket /= nb;
+  average_variation_per_bucket /= static_cast<float>(nb);
 
   TTS_LESS(average_variation_per_bucket, 0.25f);
 };
